Added Font to Score Text and Fixed Score Keeper
The score would not go past one and since the score keeper was on all three of the coin objects they were each keeping their own score. So it seemed like the score wouldn't go past one when really I was just seeing the stored score of the coins I was collecting. I had to give the score Keeper script to only one of the game objects and put triggers on the coins so that the player only sees one stored score and the coins only add to that score not their own individual scores.