Started Scripting Scoreboard
So far I have been able to add example scores to the scoreboard working with a script that duplicates a template and sorts values input into the script in the correct order. The script is not yet connected to the main game so playing the game does not effect the score board.